requirements.txt
fastapi[standard]
uvicorn
streamlit
google-genai
python-dotenv
pandas
numpy
plotly
requests
